helm/docker/requirements.txt
# Flask and web framework dependencies
Flask==2.3.3
Flask-SocketIO==5.3.6
Werkzeug==2.3.7

# Kubernetes client
kubernetes==27.2.0

# Async and threading support
eventlet==0.33.3
python-socketio==5.8.0

# YAML processing
PyYAML==6.0.1

# HTTP requests
requests==2.31.0

# Process management
psutil==5.9.5

# Logging and utilities
python-dateutil==2.8.2
